fixed unit conversion bug

+++ src/localization/src/can_sensors/imu_can_sensor.cpp
@@ -12,18 +12,20 @@ int IMUCanSensor::recieveData(void * ret)
 	if (canRecieve(recieveBuffer) ==1)
 	{
 
-		//recieved data is in the form [X LSB, X MSB, Y LSB, Y MSB, Z LSB, Z MSB]
-		//representing 16bits of signed data from -2g to +2g
 	
 		imuData *realret;
 		realret = (imuData*)ret;
 
-		//populate ret, this isn't accurate (correct) yet
-		int tempx = recieveBuffer[0] | recieveBuffer[1]<<8; //make an int out of 2 uints
+		//recieved data is in the form [X LSB, X MSB, Y LSB, Y MSB, Z LSB, Z MSB]
+		//representing 16bits of signed data from -2g to +2g
+
+
+		//populate ret, this should be correct
+		int16_t tempx = recieveBuffer[0] | recieveBuffer[1]<<8; //make an int out of 2 uints
 		realret->x = tempx/32768.0 * 2.0 * 9.8; //convert to m/s^2, tempx/max value * full scale
-		int tempy = recieveBuffer[2] | recieveBuffer[3]<<8;
+		int16_t tempy = recieveBuffer[2] | recieveBuffer[3]<<8;
 		realret->y = tempy/32768.0 * 2.0 * 9.8;
-		int tempz = recieveBuffer[4] | recieveBuffer[5]<<8;
+		int16_t tempz = recieveBuffer[4] | recieveBuffer[5]<<8;
 		realret->z = tempz/32768.0 * 2.0 * 9.8;
 
 		//return some sensible coding of how it went
